#1067 Double -> BigDecimal
git-svn-id: svn://svn.opentaps.org/opentaps_all/branches/ofbiz-0904-1.4.0-preview1@13275 d3523486-f5fe-0310-a41a-a15a4e76f3c7

+++ hot-deploy/opentaps-common/src/common/org/opentaps/common/order/shoppingcart/OpentapsShippingEstimateWrapper.java
@@ -312,7 +312,7 @@ public class OpentapsShippingEstimateWrapper implements Serializable {
         if (UtilValidate.isEmpty(storeCarrierShipMethod)) {
             return null;
         }
-        Double est = (Double) shippingEstimates.get(storeCarrierShipMethod);
+        BigDecimal est = (BigDecimal) shippingEstimates.get(storeCarrierShipMethod);
         if (UtilValidate.isEmpty(est)) {
             return null;
         }
